ChainFund-backend/ChainFund/chainfund-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
stellar-sdk==9.0.0
pydantic
python-dotenv==1.0.0
python-multipart==0.0.6
requests==2.31.0
pymongo==4.6.0
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
beanie==1.25.0
httpx==0.25.2